requirements.txt
alembic==1.0.11
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coverage==4.5.4
dominate==2.3.5
Faker==2.0.0
Flask==1.1.1
Flask-Admin==1.5.0
Flask-Bootstrap4==4.0.2
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.14
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.6
text-unidecode==1.2
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
